(chore): refactor to move tests into __tests__ folder
