+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should create a rounded avatar.snap.png
