+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should create a square avatar.snap.png
