+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should render an icon.snap.png
