+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should set a background color.snap.png
