+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should set a text color.snap.png
