+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should set the font-size.snap.png
