+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should set the size with standard size names.snap.png
