+++ ui/src/components/avatar/__tests__/All Specs/QAvatar -- should set the size.snap.png
